Fix issues mentioned in pull request

+++ Chap7/Ex1.cs
@@ -28,8 +28,8 @@ namespace Chap7 {
             return a * b;
         }
 
-        public int divide(int a, int b) {
-            return a / b;
+        public double divide(int a, int b) {
+            return (double)a / b;
         }
     }
 }
